feat(configuration): Update application configuration for AWS and database connections; refactor quantity fields to use double type

+++ warehouse-service/src/main/java/sum25/group03/warehouseservice/dto/request/SupplyReq.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class SupplyReq {
     @NotNull
     @Positive(message = "quantityReceived must be positive")
-    private BigDecimal quantityReceived;
+    private double quantityReceived;
     @NotBlank
     private String lotNumber;
     @NotBlank
