+++ warehouse-service/src/main/java/sum25/group03/warehouseservice/entity/ReagentInventory.java
@@ -26,7 +26,7 @@ public class ReagentInventory {
     private String lotNumber;
 
     @Column(name = "quantity_available", nullable = false)
-    private BigDecimal quantityAvailable;
+    private double quantityAvailable;
 
     @Column(name = "expiry_date", nullable = false)
     private LocalDate expiryDate;
